A whole lot
So I am currently working on the assumption that calling methods is possible in C++. Continuing that line of assumption I am setting up different methods for each of the things we want the motors to do. So far I have made  the base code for each of the methods. Might try making this a pull request pending an answer as to wether we can